Added suport for user-defined number of rows fetched in SQL results.
User can now specify the number of rows to retrieve for each fetch by
setting the variable 'parcelsize'.  If no value is specified, or if the
value specified is less than 1, directly use dbReadTable (which fetches all
rows at once).  This provides a considerable speed improvement for datasets
that will fit in memory.  On a table with about a quarter of a million
rows, fetch time went from approximately 90 seconds to 1.5 seconds (using
dbReadTable) compared to the previous version.